API 통해 정책DB화, 리뷰사항 반영
