+++ src/main/java/com/arom/polisee/domain/api/controller/ApiController.java
@@ -21,7 +21,8 @@ public class ApiController {
     // 정책 리스트 저장
     @GetMapping("/policies/fetch")
     public String fetchPolicies() {
-        apiService.fetchAndSavePolicies(100);
+        apiService.fetchPolicyRequirements(500);
+        apiService.fetchPolicies(500);
         return "정책 리스트 저장 완료!";
     }
