--- src/main/java/com/arom/polisee/domain/policy_requirements/dto/PolicyRequirementsDTO2.java
@@ -1,114 +1,0 @@
-package com.arom.polisee.domain.policy_requirements.dto;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.Getter;
-
-@Getter
-public class PolicyRequirementsDTO2 {
-
-    @JsonProperty("서비스ID")
-    private String id;
-
-    @JsonProperty("서비스명")
-    private String name;
-
-    @JsonProperty("JA0101")
-    private String JA0101;
-
-    @JsonProperty("JA0102")
-    private String JA0102;
-
-    @JsonProperty("JA0110")
-    private String JA0110;
-
-    @JsonProperty("JA0111")
-    private String JA0111;
-
-    @JsonProperty("JA0201")
-    private String JA0201;
-    @JsonProperty("JA0202")
-    private String JA0202;
-    @JsonProperty("JA0203")
-    private String JA0203;
-    @JsonProperty("JA0204")
-    private String JA0204;
-    @JsonProperty("JA0205")
-    private String JA0205;
-    @JsonProperty("JA0301")
-    private String JA0301;
-    @JsonProperty("JA0302")
-    private String JA0302;
-    @JsonProperty("JA0303")
-    private String JA0303;
-    @JsonProperty("JA0313")
-    private String JA0313;
-    @JsonProperty("JA0314")
-    private String JA0314;
-    @JsonProperty("JA0315")
-    private String JA0315;
-
-    @JsonProperty("JA0316")
-    private String JA0316;
-
-    @JsonProperty("JA0317")
-    private String JA0317;
-
-    @JsonProperty("JA0318")
-    private String JA0318;
-
-    @JsonProperty("JA0319")
-    private String JA0319;
-
-    @JsonProperty("JA0320")
-    private String JA0320;
-
-    @JsonProperty("JA0322")
-    private String JA0322;
-
-    @JsonProperty("JA0326")
-    private String JA0326;
-
-    @JsonProperty("JA0327")
-    private String JA0327;
-
-    @JsonProperty("JA0328")
-    private String JA0328;
-
-    @JsonProperty("JA0329")
-    private String JA0329;
-
-    @JsonProperty("JA0330")
-    private String JA0330;
-
-    @JsonProperty("JA0401")
-    private String JA0401;
-
-    @JsonProperty("JA0402")
-    private String JA0402;
-
-    @JsonProperty("JA0403")
-    private String JA0403;
-
-    @JsonProperty("JA0404")
-    private String JA0404;
-
-    @JsonProperty("JA0410")
-    private String JA0410;
-
-    @JsonProperty("JA0411")
-    private String JA0411;
-
-    @JsonProperty("JA0412")
-    private String JA0412;
-
-    @JsonProperty("JA0413")
-    private String JA0413;
-
-    @JsonProperty("JA0414")
-    private String JA0414;
-
-    @JsonProperty("JA1201")
-    private String JA1201;
-
-}
-
