+++ src/main/java/com/arom/polisee/domain/policy_requirements/repository/PolicyRequirementsRepository.java
@@ -1,13 +1,17 @@
 package com.arom.polisee.domain.policy_requirements.repository;
 
 import com.arom.polisee.domain.policy_requirements.entity.PolicyRequirements;
+import com.nimbusds.openid.connect.sdk.assurance.Policy;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
+import java.util.Optional;
 
 @Repository
+@Slf4j
 public class PolicyRequirementsRepository {
 
     @PersistenceContext
@@ -21,9 +25,8 @@ public class PolicyRequirementsRepository {
         for (PolicyRequirements policyRequirement : policyRequirements) {
             save(policyRequirement);
         }
-        em.flush();
     }
-    public PolicyRequirements findById(String id) {
-        return em.find(PolicyRequirements.class, id);
+    public Optional<PolicyRequirements> findById(String id) {
+        return Optional.ofNullable(em.find(PolicyRequirements.class, id));
     }
 }
