+++ src/main/resources/application.yml
@@ -11,7 +11,7 @@ spring:
     hibernate:
       ddl-auto: create-drop
     generate-ddl: true
-    show-sql: true
+    show-sql: false
     open-in-view: false
     properties:
       hibernate:
